requirements.txt
fastapi
waitress
mysql-connector-python
python-jose
passlib[bcrypt]
requests
beautifulsoup4
reportlab
python-dotenv

